scripts/requirements.txt
cryptography
python-jenkins
pyyaml
requests
